App.js разделено по папкам components и hooks
